refs #1804 Move user timeline under nested directory
